Add pre-commit
Adds `pre-commit` to the template.

Swaps `tox` for `nox`

Adds `poetry`

Moves `mypy` config to `pyproject.toml`

Adds `py.typed` to the package, to carry forward type-hints.

Added typehints where required for strict checking to pass.

Cleans up the `inv clean` command.

These were a couple of big changes, primarily motivated by the need to
get proper locking in place and use the same versions of the
dependencies in the tests as are used during runtime.

+++ {{ cookiecutter.project_name }}/tasks.py
@@ -1,6 +1,5 @@
 """Development tasks for {{ cookiecutter.project_name }}."""
 import os
-from glob import glob
 from pathlib import Path
 from shlex import quote
 from shutil import rmtree
@@ -9,11 +8,9 @@ from tempfile import TemporaryDirectory
 from invoke import Collection, task
 
 # Change the CWD to the repo root.
-_LAST_DIR = None
 while not Path("./tasks.py").exists():
     os.chdir("..")
-    _CURRENT_DIR = Path(".").resolve()
-    if _CURRENT_DIR == _LAST_DIR:
+    if Path(".").resolve() == Path(".").resolve().parent:
         # We hit the FS root :(
         raise FileNotFoundError("Could not find the repository root.")
 
@@ -27,78 +24,77 @@ def echo(msg):
     print(ECHO_PREFIX + str(msg))
 
 
-@task(name="pindeps")
-def pindeps(ctx, generate_hashes=True, upgrade=True):
-    """Pin dependencies."""
-    argv = ["python", "-m", "piptools", "compile"]
-    if generate_hashes:
-        argv.append("--generate-hashes")
-    if upgrade:
-        argv.append("-U")
-    os.environ["CUSTOM_COMPILE_COMMAND"] = "inv[oke] pindeps"
-    argv.append("setup.cfg")
-    ctx.run(" ".join([quote(x) for x in argv]))
-    req_path = Path(".") / "requirements.txt"
-    req_path = req_path.resolve()
-    echo(f"Dependencies pinned in {str(req_path)}")
+echo(f"Current Working Directory: {str(Path('.').resolve())}")
 
 
-@task(name="black")
-def run_black(ctx, warn=False):
-    """Run `black` to autoformat the source code."""
-    # The result of this command should satisfy the
-    # corresponding tox testenv check_black
-    ctx.run("python -m black ./src ./tests ./tasks.py", warn=warn)
+def run_argv(ctx, argv, warn=False):
+    """Run a constructed argv via the provided context."""
+    return ctx.run(" ".join([quote(x) for x in argv]), warn=warn)
 
 
-@task(name="blacken_docs")
-def run_blacken_docs(ctx, warn=False):
-    """Run `blacken-docs` to autoformat code in the documentation."""
-    # The result of this command should satisfy the
-    # corresponding tox testenv check_blacken_docs
-    docs_paths = ["README.md"]
-    docs_paths += glob("docs/**/*.rst", recursive=True)
-    docs_paths += glob("src/**/*.py", recursive=True)
-    docs_paths = [quote(p) for p in docs_paths]
-    ctx.run("blacken-docs %s" % " ".join(docs_paths), warn=warn)
-
-
-@task(name="isort")
-def run_isort(ctx, warn=False):
-    """Run `isort` to autoformat the source code."""
-    # The result of this command should satisfy the
-    # corresponding tox testenv check_isort
-    ctx.run("python -m isort ./src ./tests ./tasks.py", warn=warn)
+@task(name="pre-commit")
+def run_pre_commit(ctx, hook=None, warn=False):
+    """Run pre-commit against all files."""
+    argv = [
+        "poetry",
+        "run",
+        "pre-commit",
+        "run",
+        "--all-files",
+        "--show-diff-on-failure",
+    ]
+    if hook:
+        argv.append(hook)
+    run_argv(ctx, argv, warn=warn)
 
 
 @task(name="autoformatters")
-def run_autoformatters(ctx, warn=False):
+def run_autoformatters(ctx, warn=True):
     """Run all the autoformatters."""
-    run_black(ctx, warn=warn)
-    run_blacken_docs(ctx, warn=warn)
-    run_isort(ctx, warn=warn)
+    run_pre_commit(ctx, hook="black", warn=warn)
+    run_pre_commit(ctx, hook="blacken-docs", warn=warn)
+    run_pre_commit(ctx, hook="isort", warn=warn)
+    run_pre_commit(ctx, hook="end-of-file-fixer", warn=warn)
+    run_pre_commit(ctx, hook="trailing-whitespace", warn=warn)
+
+
+@task(name="nox", iterable=["env"])
+def run_nox(ctx, warn=False, env=None):
+    """Run nox."""
+    argv = ["poetry", "run", "nox"]
+    if env:
+        for environment in env:
+            argv.extend(["-e", environment])
+
+    run_argv(ctx, argv, warn=warn)
 
 
 @task(name="tests", iterable=["env"])
-def run_tests(ctx, autoformat=True, warn=False, env=None, recreate=False):
+def run_tests(ctx, autoformat=True, env=None, warn=False):
     """Run the tests."""
     if autoformat:
+        echo("Running autoformatters...")
         run_autoformatters(ctx, warn=warn)
-    argv = ["tox"]
-    if env:
-        argv.append("-e")
-        argv.append(",".join(env))
-    if recreate:
-        argv.append("--recreate")
-    ctx.run(" ".join([quote(x) for x in argv]), warn=warn)
+    clean_coverage()
+    run_nox(ctx, warn=warn, env=env)
 
 
 @task(name="docs")
-def build_docs(ctx, clean=True, buildername="html"):
+def build_docs(ctx, clean_=True, buildername="html"):
     """Build the documentation."""
-    if clean:
-        clean_docs(ctx)
-    ctx.run(f"sphinx-build docs docs_out --color -W -b{buildername}")
+    if clean_:
+        clean_docs()
+    argv = [
+        "poetry",
+        "run",
+        "sphinx-build",
+        "docs",
+        "docs_out",
+        "--color",
+        "-W",
+        f"-b{buildername}",
+    ]
+    run_argv(ctx, argv)
     docs_dir = Path(".") / "docs_out"
     docs_dir = docs_dir.resolve()
     echo(f"Docs available in {str(docs_dir)}")
@@ -119,62 +115,60 @@ def remove_directory(directory, raise_=False):
     echo(f"{str(directory)} removed.")
 
 
-@task(name="dists")
-def clean_dists(ctx):  # pylint:disable=W0613
+def clean_dists():
     """Remove existing distributions."""
     dist_dir = Path("./dist").resolve()
     remove_directory(dist_dir)
 
 
-@task(name="docs")
-def clean_docs(ctx):  # pylint:disable=W0613
+def clean_docs():
     """Remove existing docs."""
     docs_out_dir = Path("./docs_out").resolve()
     remove_directory(docs_out_dir)
 
 
-@task(name="compiled")
-def clean_compiled(ctx):  # pylint:disable=W0613
-    """Remove compilation artifacts."""
+def clean_compiled(dir_=None, _recursing=False):
+    """Remove compiled python artifacts from a directory recursively."""
+    if dir_ is None:
+        dir_ = Path().resolve()  # cwd
 
-    def clean_dir(dir_):
-        """Remove compiled python artifacts from a directory."""
-        for entry in dir_.iterdir():
-            if entry.name.endswith(".pyc"):
-                entry.unlink()
-                continue
-            if entry.name == "__pycache__" or entry.name.endswith(".egg-info"):
-                rmtree(entry)
-                continue
-            if entry.is_dir():
-                clean_dir(entry)
-
-    dirs_to_clean = [
-        Path("./src").resolve(),
-        Path("./tests").resolve(),
+    # Names to skip removing or recursing into
+    bad_names = [
+        "venv",
+        ".venv",
+        "env",
+        ".env",
+        ".git",
     ]
-    for dir_ in dirs_to_clean:
-        echo(f"Removing compiled/cache artifacts in {str(dir_)}")
-        clean_dir(dir_)
-    echo("Compiled artifacts and caches removed")
+
+    if not _recursing:
+        echo(f"Removing compiled artifacts from {str(dir_)} recursively...")
+    for entry in dir_.iterdir():
+        if entry.name in bad_names:
+            continue
+        if entry.name.endswith(".pyc"):
+            entry.unlink()
+            continue
+        if entry.name == "__pycache__" or entry.name.endswith(".egg-info"):
+            rmtree(entry)
+            continue
+        if entry.is_dir():
+            clean_compiled(dir_=entry, _recursing=True)
 
 
-@task(name="tox")
-def clean_tox(ctx):  # pylint:disable=W0613
-    """Remove the .tox cache directory."""
-    tox_dir = Path("./.tox").resolve()
-    remove_directory(tox_dir)
+def clean_nox():
+    """Remove the .nox cache directory."""
+    nox_dir = Path("./.nox").resolve()
+    remove_directory(nox_dir)
 
 
-@task(name="mypy")
-def clean_mypy(ctx):  # pylint:disable=W0613
+def clean_mypy():
     """Remove the .mypy_cache directory."""
     mypy_cache_dir = Path("./.mypy_cache").resolve()
     remove_directory(mypy_cache_dir)
 
 
-@task(name="coverage")
-def clean_coverage(ctx):  # pylint:disable=W0613
+def clean_coverage():
     """Remove the .coverage cache."""
     coverage_cache_file = Path("./.coverage").resolve()
     if coverage_cache_file.exists() and coverage_cache_file.is_file():
@@ -184,68 +178,68 @@ def clean_coverage(ctx):  # pylint:disable=W0613
         echo(f"No file found at {str(coverage_cache_file)}")
 
 
-@task(name="wheelhouse")
-def clean_wheelhouse(ctx):  # pylint:disable=W0613
-    """Remove the wheelhouse."""
-    wheelhouse_dir = Path("./wheelhouse").resolve()
-    remove_directory(wheelhouse_dir)
-
-
-@task(name="coverage-report")
-def clean_coverage_report(ctx):  # pylint:disable=W0613
+def clean_coverage_report():
     """Remove the html coverage report."""
     coverage_report_dir = Path("./htmlcov").resolve()
     remove_directory(coverage_report_dir)
 
 
-@task(name="build-dir")
-def clean_build_dir(ctx):  # pylint:disable=W0613
+def clean_build_dir():
     """Remove the build directory."""
     build_directory = Path("./build").resolve()
     remove_directory(build_directory)
 
 
-@task(name="all")
-def clean_all(
+def clean_pytest():
+    """Remove the pytest cache directory."""
+    pytest_cache_directory = Path("./.pytest_cache").resolve()
+    remove_directory(pytest_cache_directory)
+
+
+@task
+def clean(
     ctx,
     dists=True,
     docs=True,
     compiled=True,
-    tox=True,
+    nox=True,
     mypy=True,
+    pytest=True,
     coverage=True,
-    wheelhouse=True,
     coverage_report=True,
     build_directory=True,
-):  # pylint:disable=R0913
-    """Remove all clean-able artifacts."""
-    if dists:
-        clean_dists(ctx)
-    if docs:
-        clean_docs(ctx)
-    if compiled:
-        clean_compiled(ctx)
-    if tox:
-        clean_tox(ctx)
-    if mypy:
-        clean_mypy(ctx)
-    if coverage:
-        clean_coverage(ctx)
-    if wheelhouse:
-        clean_wheelhouse(ctx)
-    if coverage_report:
-        clean_coverage_report(ctx)
-    if build_directory:
-        clean_build_dir(ctx)
+):  # pylint:disable=W0613,R0913
+    """
+    Clean up all caches and generated artifacts.
+
+    Note that the default is to clean _everything_, and if you would like to
+    preserve any artifacts/caches you should pass the corresponding `--no-...`
+    flag to this command.
+    """
+    cleaners = (
+        (dists, clean_dists),
+        (docs, clean_docs),
+        (compiled, clean_compiled),
+        (nox, clean_nox),
+        (mypy, clean_mypy),
+        (pytest, clean_pytest),
+        (coverage, clean_coverage),
+        (coverage_report, clean_coverage_report),
+        (build_directory, clean_build_dir),
+    )
+    for entry in cleaners:
+        if entry[0]:  # If the arg is true
+            entry[1]()  # Run the cleaner
 
 
 @task(name="dists")
-def build_dists(ctx, clean=True):
+def build_dists(ctx, clean_=True):
     """Build distribution artifacts."""
-    if clean:
-        clean_dists(ctx)
-        clean_build_dir(ctx)
-    ctx.run("python -m build -s -w .")
+    if clean_:
+        clean_dists()
+        clean_build_dir()
+    argv = ["poetry", "run", "python", "-m", "build", "."]
+    run_argv(ctx, argv)
     echo(f"Dists now available in {Path('./dist').resolve()}")
 
 
@@ -254,14 +248,10 @@ def build_zipapp(
     ctx,
     command_name=None,
     executable_name=None,
-    pin_dependencies=True,
     shebang=None,
     compress=False,
 ):  # pylint:disable=R0913
-    """Use `shiv` to produce a zipapp that includes the depndencies."""
-    if pin_dependencies:
-        pindeps(ctx)
-
+    """Use `shiv` to produce a zipapp that includes the dependencies."""
     # Defaults
     if shebang is None:
         shebang = "/usr/bin/env python3"
@@ -279,14 +269,45 @@ def build_zipapp(
     dist_dir.mkdir(exist_ok=True)
     output_fname = dist_dir / executable_name
 
-    # Run shiv
     echo("Creating zipapp with shiv")
     with TemporaryDirectory() as tmp_dir:
+        # Build a requirements.txt
+        reqs_filepath = Path(tmp_dir) / "requirements.txt"
+        echo(f"Creating requirements.txt at {str(reqs_filepath.resolve())}")
+        run_argv(
+            ctx,
+            [
+                "poetry",
+                "export",
+                "--format=requirements.txt",
+                f"--output={str(reqs_filepath.resolve())}",
+            ],
+        )
+
+        # Build a site-packages dir
+        echo(f"Creating temporary site_packages at {tmp_dir}/site-packages")
+        run_argv(
+            ctx,
+            [
+                "python",
+                "-m",
+                "pip",
+                "install",
+                "-r",
+                f"{str(reqs_filepath.resolve())}",
+                "--target",
+                f"{tmp_dir}/site-packages",
+            ],
+        )
+        ctx.run(f"python -m pip install . --target {tmp_dir}/site-packages")
+
         # Build the shiv argv
         shiv_cmd = [
+            "poetry",
+            "run",
             "shiv",
             "--site-packages",
-            tmp_dir,
+            f"{tmp_dir}/site-packages",
             "-p",
             shebang,
             "-o",
@@ -297,71 +318,63 @@ def build_zipapp(
             shiv_cmd.extend(["-c", command_name])
         if compress:
             shiv_cmd.append("--compressed")
-        echo(f"Creating temporary site_packages at {tmp_dir}")
-        # pip-sync doesn't appear to suport the `--target` option :(
-        # https://github.com/jazzband/pip-tools/issues/587
-        # Should be okay, as there aren't going to be any extra
-        # packages in the purpose build site-packages anyways.
-        ctx.run(f"python -m pip install -r requirements.txt --target {tmp_dir}")
-        ctx.run(f"python -m pip install . --target {tmp_dir}")
-        ctx.run(" ".join([quote(x) for x in shiv_cmd]))
+        # Run the shiv command
+        run_argv(ctx, shiv_cmd)
     echo(f"Created zipapp: {output_fname}")
 
 
-@task(name="wheelhouse")
-def build_wheelhouse(ctx, clean=True, pin_dependencies=True):
-    """Build a dependency wheelhouse."""
-    if clean:
-        clean_wheelhouse(ctx)
-    if pin_dependencies:
-        pindeps(ctx)
-    ctx.run("python -m pip wheel -w wheelhouse -r requirements.txt")
-    echo(f"Wheelhouse now available in {Path('./wheelhouse').resolve()}")
-
-
 @task(name="coverage-report")
-def build_coverage_report(ctx, clean=True, test=True):
+def build_coverage_report(ctx, clean_=True, test=True):
     """Build an HTML coverage report."""
-    if clean:
-        clean_coverage_report(ctx)
+    if clean_:
+        clean_coverage_report()
     if test:
         run_tests(ctx, warn=True)
-    ctx.run("coverage html")
+    run_argv(ctx, ["poetry", "run", "coverage", "html"])
     echo(f"Coverage report available at {str(Path('./htmlcov').resolve())}")
 
 
 @task(name="todos")
 def check_todos(ctx):
     """Check for `#TODO` comments in the code."""
-    ctx.run("python -m pylint --disable=all --enable=W0511 src tests")
+    argv = [
+        "poetry",
+        "run",
+        "pylint",
+        "--disable=all",
+        "--enable=W0511",
+        "src",
+        "tests",
+    ]
+    run_argv(ctx, argv)
 
 
 @task()
 def release(
-    ctx, prod=False, clean=True, build=True, skip_tests=False, autoformat=True
+    ctx, prod=False, clean_=True, build=True, skip_tests=False, autoformat=True
 ):  # pylint:disable=R0913
     """Perform a release to pypi."""
     if not skip_tests:
         run_tests(ctx, autoformat=autoformat)
-    if clean:
-        clean_dists(ctx)
+    if clean_:
+        clean_dists()
     if build:
         build_dists(ctx)
-    cmd = "python -m twine upload"
+    argv = ["python", "-m", "twine", "upload"]
     if not prod:
         echo("Uploading to test.pypi")
-        cmd += " --repository-url https://test.pypi.org/legacy"
+        argv.extend(["--repository-url", "https://test.pypi.org/legacy"])
     else:
         echo("--prod flag present. Uploading to pypi")
-    cmd += " ./dist/*"
-    ctx.run(cmd)
+    argv.append("./dist/*")
+    run_argv(ctx, argv)
     echo("Upload complete")
 
 
 @task(name="docs", iterable=["additional_dir"])
 def serve_docs(ctx, todos=False, additional_dir=None, open_browser=False):
     """Serve the docs on localhost:8000. Reload on changes."""
-    argv = ["sphinx-autobuild"]
+    argv = ["poetry", "run", "sphinx-autobuild"]
 
     if open_browser:
         argv.append("--open-browser")
@@ -385,7 +398,7 @@ def serve_docs(ctx, todos=False, additional_dir=None, open_browser=False):
     if todos:
         os.environ["SPHINX_DISPLAY_TODOS"] = "true"
 
-    ctx.run(" ".join([quote(x) for x in argv]))
+    run_argv(ctx, argv)
 
 
 # Make implicit root explicit
@@ -395,16 +408,15 @@ ns.configure({"run": {"pty": True, "echo": True}})
 
 
 # Root commands
-ns.add_task(pindeps)
 ns.add_task(release)
+ns.add_task(clean)
 
 
 # Define the "run" subcommand
 run_ns = Collection("run")
 run_ns.add_task(run_autoformatters)
-run_ns.add_task(run_black)
-run_ns.add_task(run_blacken_docs)
-run_ns.add_task(run_isort)
+run_ns.add_task(run_nox)
+run_ns.add_task(run_pre_commit)
 run_ns.add_task(run_tests)
 
 
@@ -412,22 +424,8 @@ run_ns.add_task(run_tests)
 build_ns = Collection("build")
 build_ns.add_task(build_dists, default=True)
 build_ns.add_task(build_docs)
-build_ns.add_task(build_wheelhouse)
 build_ns.add_task(build_coverage_report)
 build_ns.add_task(build_zipapp)
-
-
-# Define the "clean" subcommand
-clean_ns = Collection("clean")
-clean_ns.add_task(clean_all, default=True)
-clean_ns.add_task(clean_dists)
-clean_ns.add_task(clean_docs)
-clean_ns.add_task(clean_compiled)
-clean_ns.add_task(clean_tox)
-clean_ns.add_task(clean_mypy)
-clean_ns.add_task(clean_coverage)
-clean_ns.add_task(clean_wheelhouse)
-clean_ns.add_task(clean_coverage_report)
 
 
 # Define the "check" subcommand
@@ -443,6 +441,5 @@ serve_ns.add_task(serve_docs)
 # Add custom subcommands to root namespace
 ns.add_collection(run_ns)
 ns.add_collection(build_ns)
-ns.add_collection(clean_ns)
 ns.add_collection(check_ns)
 ns.add_collection(serve_ns)
